did some changes

+++ src/functions/firestore.js
@@ -1,3 +1,5 @@
+import { useEffect , useState } from "react"
+
 import { firestore } from "../firebase";
 import store from "../redux/configureStore";
 import { 
@@ -10,9 +12,10 @@ import {
     setDoc,
     addDoc,
     serverTimestamp,
-    where , query,
+    where , query, limit,
     deleteDoc, onSnapshot,
-    updateDoc
+    updateDoc,
+    orderBy , getCountFromServer
 } from "firebase/firestore";
 
 // collections
@@ -54,24 +57,15 @@ export const getUserDoc = async ({id , email , displayName}) => {
     getUserDoc({id , email , displayName});
 }
 
-export const addRequest = async (uid,email) => {
-    try{
-        await setDoc(doc(requests , uid),{
-            email,
-            uid
-        })
-        window.location.reload();
-    }
-    catch(err){
-        console.log(err)
-        return false
-    }
-}
+const check = async (id) => {
 
-export const checkRequest = async (id) => {
+    const {admin , uid } = store?.getState().authSlice.user;
+
     try{
-        const snapshot = await getDoc(doc(requests , id))
-        return snapshot.exists()
+        const snapshot = await getDoc(doc(books , id));
+        const book = snapshot.data()
+        if(book.createdBy !== uid && !admin) return false
+        return true
     }
     catch(err){
         console.log(err)
@@ -96,7 +90,8 @@ export const createBookDoc = async ({
         cover,
         categorys,
         createdAt : serverTimestamp(),
-        createdBy : store.getState().authSlice.user.uid
+        createdBy : store.getState().authSlice.user.uid,
+        likedBy : []
     })
     .then(() => {
         setSuccess("موفق")
@@ -107,9 +102,10 @@ export const createBookDoc = async ({
 }
 
 export const getBooksByQuery = async (
-    property , operator , value, setData , setPending , setError
+    property , operator , value, setData , setPending , setError,
+    LIMIT
     ) => {
-    const q = query(books , where(property , operator , value));
+    const q = query(books , where(property , operator , value) ,limit(LIMIT));
 
     const unsubscribe = onSnapshot(q , (snapshots) => {
             setData([])
@@ -125,7 +121,8 @@ export const getBooksByQuery = async (
             const finale = snapshots.docs.map((item) => {
                 return {
                     ...item.data(),
-                    docId : item.id
+                    docId : item.id,
+                    createdAt : item.createdAt
                 }
             }) 
             setData(finale)
@@ -151,6 +148,10 @@ export const getBook = async (id) => {
 }
 
 export const deleteBook = async (id , setId) => {
+
+    const available = await check(id)
+    if(!available) return
+    
     try{
         await deleteDoc(doc(books,id))
         setId("")
@@ -161,6 +162,10 @@ export const deleteBook = async (id , setId) => {
 }
 
 export const updateBook = async (id , data) => {
+
+    const available = await check(id)
+    if(!available) return
+
     try{
         await updateDoc(doc(books,id),
             {
@@ -176,4 +181,112 @@ export const updateBook = async (id , data) => {
     catch(err){
         console.log(err.toString())
     }
+}
+
+export const useWatchForBook = (id) => {
+
+    const [data , setData] = useState(null);
+    const [pending , setPending] = useState(null);
+    const [error , setError] = useState(null);
+
+    useEffect(() => {
+
+        if(!id) return
+        setData(null)
+        setPending(true)
+        setError(null)
+
+        const unsub = onSnapshot(doc(books , id),(snapshot) => {
+            try{
+                const book = {
+                    ...snapshot.data(),
+                    docId : snapshot.id
+                }
+
+                setData(book)
+                setPending(false)
+                setError(null)
+            }
+            catch(err){
+                setData(null)
+                setPending(false)
+                setError(err.name)
+            }
+        })
+
+        return () => unsub()
+    },[id])
+
+    return[data , pending , error]
+}
+
+export const useGetBooksByQuery = (property , operator , value , LIMIT) => {
+    const [data , setData] = useState([]);
+    const [pending , setPending] = useState(true);
+    const [error , setError] = useState(null);
+
+    useEffect(() => {
+
+        const q = query(books , where(property , operator , value) ,limit(LIMIT));
+
+        const unsub = onSnapshot(q , (snapshots) => {
+            setData([])
+            setPending(false)
+            setError(null)
+            
+            if(!snapshots.docs[0]){
+                setData([])
+                setPending(false)
+                setError(null)
+                return
+            } 
+            const finale = snapshots.docs.map((item) => {
+                return {
+                    ...item.data(),
+                    docId : item.id,
+                    createdAt : item.createdAt
+                }
+            }) 
+            setData(finale)
+            setPending(false)
+            setError(null)
+        });        
+
+        return () => unsub()
+    },[property , operator , value , LIMIT])
+
+    return [ data , pending , error ];
+} 
+
+export const likeTheBook = async (docId , uid) => {
+    try{
+        const docRef = doc(books,docId)
+        const snapshot = await getDoc(docRef)
+        const likedBy = snapshot.data().likedBy
+        if(likedBy.includes(uid)){
+            let i = likedBy.findIndex(item => item === uid);
+            likedBy.splice(i,1)
+        }
+        else{
+            likedBy.push(uid);
+        }
+        updateDoc(docRef , {
+            likedBy
+        })
+    }
+    catch(err){
+        console.log(err.name)
+    }
+}
+
+export const useGetDataWithPagination = () => {
+    const dataLimit = 50;
+    const [data,setData] = useState([]);
+    const [pending,setPending] = useState(true);
+    const [error,setError] = useState([]);
+    const [hasMore,setHasMore] = useState(false);
+
+    useEffect(() => {
+        const q = query(where("name"))
+    },[])
 }
